ISSUE-#5: refactor/move_config_module コンフィグ設定モジュールの
リファクタリングによるファイル移動

confモジュールをインフラパッケージ移動
importパス修正
