+++ trend-collector/trendcollector/libs/__init__.py
@@ -2,6 +2,6 @@
 
 from .services.client import *
 from .infrastractures.twitter_v2 import *
-from .conf import *
+from .infrastractures.conf import *
 from .response import *
 from .infrastractures import *
