+++ trend-collector/trendcollector/libs/infrastractures/__init__.py
@@ -1,3 +1,4 @@
+from .conf import *
 from .twitter_account_repository import *
 from .trend_repository import TrendRepository
 from .schemas import Base
