+++ trend-collector/trendcollector/libs/infrastractures/conf.py
